Server Login 로직 변경

+++ core/network/src/main/java/com/hmh/hamyeonham/core/network/login/model/LoginResponse.kt
@@ -9,7 +9,7 @@ data class LoginResponse(
     @SerialName("token")
     val token: Token? = null,
     @SerialName("userId")
-    val userId: Int? = null
+    val userId: Long? = null
 ) {
     @Serializable
     data class Token(
